requirements_dev.txt
-r requirements.txt
autoflake==1.4
autopep8==1.6.0
black==22.3.0
coverage==6.4
flake8==4.0.1
isort==5.10.1
pylint==2.14.0
pylint-django==2.5.3
# tblib==1.7.0
